Resolves #4; Finished configuring cmake to use fetchContent to download, build, and link our dependencies
