+++ .gitignore
@@ -1,8 +1,22 @@
-build/*
-bin/*
 assets/wee
 assets/tiled
 
+imgui.ini
+
+bin/*
+build/*
+**/cmake-build-debug
+**/CMakeCache.txt
+**/cmake_install.cmake
+**/install_manifest.txt
+**/CMakeFiles/
+**/CTestTestfile.cmake
+**/Makefile
+**/*.cbp
+**/CMakeScripts
+**/compile_commands.json
+
+*.plist
 *.vim
 *.pdf
 *.png
