+++ CMakeLists.txt
@@ -1,8 +1,9 @@
 cmake_minimum_required(VERSION 3.16)
-project(ComputerGame VERSION 0.0.0 LANGUAGES CXX)
+project(ComputerGame
+  VERSION 0.0.0
+  LANGUAGES CXX
+)
 
-set(CMAKE_C_COMPILER "/usr/bin/clang")
-set(CMAKE_CXX_COMPILER "/usr/bin/clang++")
 set(CMAKE_CXX_STANDARD 17)
 SET(CMAKE_CXX_STANDARD_REQUIRED ON)
 SET(CMAKE_EXPORT_COMPILE_COMMANDS ON)
@@ -18,10 +19,6 @@ message("C++ compiler flags: ${CMAKE_CXX_FLAGS}")
 # Libraries
 SET_PROPERTY(GLOBAL PROPERTY TARGET_SUPPORTS_SHARED_LIBS TRUE)
 
-find_package(SFML 2.5 REQUIRED COMPONENTS system window graphics)
-set(SFML_LIBRARIES sfml-system sfml-window sfml-graphics)
 
-find_package(Threads REQUIRED)
-# find_package(Boost 1.75.0 COMPONENTS log REQUIRED)
-
+add_subdirectory(vendor)
 add_subdirectory(src)
