+++ src/Engine/CMakeLists.txt
@@ -12,4 +12,4 @@ add_library(engine
 
 target_include_directories(engine PUBLIC "${CMAKE_SOURCE_DIR}/src")
 
-target_link_libraries(engine traits ${SFML_LIBRARIES})
+target_link_libraries(engine traits ImGui-SFML::ImGui-SFML)
