+++ src/Game/CMakeLists.txt
@@ -6,4 +6,4 @@ add_library(game
 
 target_include_directories(game PUBLIC "${CMAKE_SOURCE_DIR}/src")
 
-target_link_libraries(game traits engine ${SFML_LIBRARIES})
+target_link_libraries(game traits engine ImGui-SFML::ImGui-SFML)
