+++ src/Traits/CMakeLists.txt
@@ -7,6 +7,9 @@ add_library(traits
 )
 
 # Includes in other places just have to reference the name of the file since they are so common
-target_include_directories(traits PUBLIC "${CMAKE_SOURCE_DIR}/src/Traits" "${CMAKE_SOURCE_DIR}/vendor/spdlog/include/")
+target_include_directories(traits PUBLIC
+  "${CMAKE_SOURCE_DIR}/src/Traits"
+  "${SPDLOG_DIR}/include/"
+)
 
-target_link_libraries(traits ${SFML_LIBRARIES})
+target_link_libraries(traits ImGui-SFML::ImGui-SFML)
